Reload the whole page in order to avoid cached dynamic imports
JS's dynamic import will not reload already loaded modules, but we want
to reload and iterate quickly. WebAssembly module can be reloaded explicitly,
but it won't work for JS modules unfortunately. So an easy way to ensure
the module is always reloaded is to reload the whole page.

Of course, if the server caches by duration, it may not be reloaded, but
it's not our HTML's problem.