Somewhat breaking change: trailing semicolon drops value instead of just optional
It is the same semantics as Rust. The semicolon is no longer simply optional
but meaningful. If you want to return a value, you need to omit the semicolon.

It makes a one-linear function a bit counterintuitive in that you can't put
a semicolon if you omit the brace, e.g.

    let dot(lhs: Vec2, rhs: Vec2) = lhs.x * rhs.x + lhs.y * rhs.y

I'm not the biggest fan of this syntax, but I don't have much better solution either.